fixes #4605
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/simpleshoutbox/trunk@24324 735d13b6-9817-0410-8766-e36946ffe9aa

+++ ext_tables.sql
@@ -6,7 +6,7 @@ CREATE TABLE tx_simpleshoutbox_messages (
 	tstamp int(11) NOT NULL default '0',
 	crdate int(11) NOT NULL default '0',
 	deleted tinyint(4) NOT NULL default '0',
-	userid tinyint(4) default '0',
+	userid int(11) default '0',
 	`name` tinytext,
 	message tinytext,
 	doublecheck tinytext,
